% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{altcpueplot}
\alias{altcpueplot}
\title{altcpueplot plots the fit to cpue and residuals for all fleets}
\usage{
altcpueplot(cpue, analysis, rundir, height = 8, CI = TRUE, console = TRUE)
}
\arguments{
\item{cpue}{the plotreport$cpue object from r4ss}

\item{analysis}{the name of the model scenario}

\item{rundir}{the directory into which to place the plot if saved}

\item{height}{the height of the plot, the default = 8, which suits 4 fleets.
if you have fewer then adjust this appropriately. If more than 5 then
external to the function it may be best to subset the cpue input data
and annotate the analysis argument to keep each plot separate.}

\item{CI}{default = TRUE confidence bounds will be plotted.}

\item{console}{default = TRUE so plot goes to console.}
}
\value{
invisibly the filename, it also plots a graph
}
\description{
altcpueplot generates a pair of plots for each fleet with cpue
data. The first is observed vs expected (the fit) and the second is an
explicit residual plot on a linear scale.
}
\examples{
# syntax: 
# extradir <- pathtopath(destination,"extra/")
# filename <- altcpueplot(plotreport$cpue,analysis="SGBC-5-4-8-6",
#                         rundir=extradir,height=8,CI=TRUE,console=TRUE)
}
