% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.r
\name{changeCtl}
\alias{changeCtl}
\title{changeCtl alters a parameter within the SS3 ctl file}
\usage{
changeCtl(
  param,
  newvalue,
  ctlfile = "ss.ctl",
  directory = NA,
  pos = 3,
  pos2 = 7
)
}
\arguments{
\item{param}{the text used to identify the line within the control file
containing the parameter to change}

\item{newvalue}{the value of the parameter to be set}

\item{ctlfile}{the name of the control file to change, defaults = "ss.ctl"}

\item{directory}{the directory in which the control exists that is to be
changed; defaults to NA but should be calc, defined as the subdirectory
in which all calculations occur within rforSS3}

\item{pos}{the position in the vector of numbers in the line; default 3}

\item{pos2}{position of the phase, which for a likelihood profile should
be turned negative; defaults to 7}
}
\value{
new control file
}
\description{
changeCtl alters a parameter within the SS3 ctl file
once it has been copied into the calc directory. This means the
original is never messed with. It finds the param string within
the ctl file andmodifies that line of text by replacing the
variable value at varpos (typically 3) with the newvalue. Once
completed the function outputs a new ctl file as text, which then
needs to be written back into the calc directory using the
write command
}
\examples{
\dontrun{
# typical syntax might be
varpos=3; pos2=7; profvalue[1]=0.15
control <- changeCtl("NatM_p_1_Fem_GP_1",profvalue[1],ctlfile="ss3.ctl",calc,
                      pos=varpos,pos2=phasepos)
write(control,file=paste0(calc,"ss3.ctl"))
print("See worked example in the rforSS3 vignette")
}
}
