% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.r
\name{changePar}
\alias{changePar}
\title{changePar is used to alter a parameter within ss.par}
\usage{
changePar(newvalue, directory)
}
\arguments{
\item{newvalue}{the new value for the given parameter}

\item{directory}{In which directory will the parameters be found.}
}
\value{
it re-writes the ss.par file and prints a confirmation to the console
}
\description{
changePar is used, when conducting a likelihood profile, to
change the value of a given parameter within ss.par so that the model
can be re-fitted while keeping this parameter constant. This function
may now be redundant as I believe r4ss now has a profile function?
}
