% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceoutput.r
\name{checkConverged}
\alias{checkConverged}
\title{checkConverged - checks variance adjustments}
\usage{
checkConverged(inLen, inAge, inrec, incpue)
}
\arguments{
\item{inLen}{\itemize{
\item the vector of numbers from fixLengthVar
}}

\item{inAge}{\itemize{
\item the vector of numbers from fixAgeVar
}}

\item{inrec}{\itemize{
\item the recdat matrix from fixRecruit
}}

\item{incpue}{\itemize{
\item the cedat matrix from fixIndex
}}
}
\value{
TRUE or FALSE; logical
}
\description{
checkConverged - checks variance adjustments. Takes the
outputs from fixLengthVar, fixAgeVar, fixRecruit, and fixIndex
found in balanceoutput.r and checks to see if the changes to
the vaiance adjustments are small enough to count as converged
}
\examples{
print("Need to develop an example")
}
