% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{cleanDir}
\alias{cleanDir}
\title{cleanDir - removes any previous analyses and plots stored in 'SubDir'}
\usage{
cleanDir(Dir, SubDir)
}
\arguments{
\item{Dir}{\itemize{
\item the storage directory containing the subdirectories for each of
the bridging analyses.
}}

\item{SubDir}{\itemize{
\item the bridging analysis subdirectory under consideration
}}
}
\value{
The files listed under description are removed from 'SubDir';
}
\description{
cleanDir -  removes any previous analyses and plots stored in
'SubDir'. Apart from the plot directory and contents it also removes
checkup.sso, CompReport.sso, covar.sso, CumReport.sso,
echoinput.sso, Forecast-report.sso, ParmTrace.sso, Report.sso,
SIS_table.sso, warning.sso, ss3.cor, ss3.std, ss3.rep, 1rundetails.txt,
2rundetails.txt, and 3rundetails.txt.  if the files are missing a warning
is given, which can be ignored. If they are presnet lots of TRUEs will
be printed.
}
\examples{
\dontrun{
# Dir <- store; SubDir <- "basecase"
print("An example has still to be written")
}
}
