% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{codeBlock}
\alias{codeBlock}
\title{codeBlock - delineate some comment lines ready to document code}
\usage{
codeBlock(rows = 2)
}
\arguments{
\item{rows}{\itemize{
\item the number of open rows within the code block ready for text;
}}
}
\value{
a block of hashes with space for descriptive text
}
\description{
codeBlock - To assist in breaking up R files into sections this
function generates a block of hashes which can be pasted into an R files
so that only the descriptive text need be altered
}
\examples{
\dontrun{
codeBlock()
codeBlock(6)
}
}
