% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{comparestats}
\alias{comparestats}
\title{comparestats takes in a set of plotreports/replists and compares them}
\usage{
comparestats(x)
}
\arguments{
\item{x}{a set of plotreports/replists output from getreplists}
}
\value{
a list of the answer, likes, param, and model matrices.
}
\description{
comparestats takes in a set of plotreports/replists collected
by the 'getreplists' function and generates an outstats object that
contains, for each scenario, summary outputs, the likelihoods, the
estimable parameter values, and the standard model structure.
}
\examples{
print("wait on example data")
}
\seealso{
{
\link{getreplists}
}
}
