% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{copyfiles}
\alias{copyfiles}
\title{copyfiles - copies required files from store to the calc directory}
\usage{
copyfiles(x, origin, destination)
}
\arguments{
\item{x}{\itemize{
\item the analysis to be conducted
}}

\item{origin}{\itemize{
\item the directory name as the source of the files to copy
}}

\item{destination}{\itemize{
\item the calc directory into which the files are copied
}}
}
\value{
Nothing, but it copies the files from the storage directory to
the calc directory ready to run SS3
}
\description{
copyfiles - copies required files from store to the calc
directory. An example store directory might be named 'balanced' or
'basecase' and the files named basecase.ctl, basecase.dat, etc. The
files copied are the ctl, dat, par, forcast, and starter
files. In the process of copying the analysis name ('basecase') is
replaced with SS3 ready for the SS3 software.
}
\examples{
\dontrun{
# x <- analysis; origin <- store; destination=calc
print("An example has still to be written")
# x=analysis; origin=store; destination=calc
}
}
