% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{dirExists}
\alias{dirExists}
\title{dirExists: Checks for the existence of a directory}
\usage{
dirExists(indir, create = TRUE)
}
\arguments{
\item{indir}{a character string containing the name of the directory to
be created if it does not already exist}

\item{create}{default = TRUE, should the input directory be created if it
does not already exist?}
}
\value{
a message to the screen if the directory exists or is created; if
the latter then it also creates the directory as listed in 'indir'.
}
\description{
dirExists: Creates a directory if it does not already exist
it use dir.create but avoids the warning message is one already exists
}
\examples{
\dontrun{
indirect <- getwd()
dirExists(indirect,create=FALSE)
}
}
