% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_extra.R
\name{do_extra}
\alias{do_extra}
\title{do_extra generates extra plots and tables to complement r4ss outputs}
\usage{
do_extra(
  plotreport,
  extradir,
  analysis,
  store,
  compare = NULL,
  paths = NULL,
  verbose = TRUE
)
}
\arguments{
\item{plotreport}{the output from SS_output}

\item{extradir}{the path to the extra directory into which to place the new
tables and plots}

\item{analysis}{the names of the SS3 model being explored}

\item{store}{the path to the directory within which the analysis is held}

\item{compare}{default = NULL. If multiple scenarios or analyses are to be
compared if they are all stored in the same 'store' then this should be
a character vector of which analyses to compare. If no
comparisons are to be made then leave as NULL.}

\item{paths}{If different locations are used to store the analyses to be
compared then this should be a character vector of the full paths. If no
comparisons are to be made then leave as NULL.}

\item{verbose}{Should warnings and progress be reported to the console?
default = TRUE.}
}
\value{
nothing but it does generate an array of plots and tables inserted
inot extradir
}
\description{
do_extra generates extra plots and tables relating to a given
scenario (analysis). It produces a new internal website for viewing the
new results.
}
\examples{
# syntax:
# do_extra(plotreport=plotreport,extradir=extradir,analysis=analysis,
#          store=store)
}
