% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{domeSS3}
\alias{domeSS3}
\title{domeSS3 calculates domed selectivity curve, pattern24 for size in SS33}
\usage{
domeSS3(p, bins)
}
\arguments{
\item{p}{a vector of six parameters.}

\item{bins}{a vector of the mean of nb age/size classes}
}
\value{
a vector of selectivities
}
\description{
domeSS3 uses 6 parameters and a set of mean size or age classes
to calculate a domed selectivity curve with a maximum of 1.0 (rescaling
can be done outside the function), but has parameters for the selectivity
of the initial and final size/age classes. There is an ascending limb and
a descending limb with the potential of a plateau in between. The
description in the SS3 technical description, p 9-10 Methot and Wetzel
(2013, supplementary material = Appendix A) is somewhat confusing, there
is a peak2 but no peak1. the sequence of calculation needs to be peak2,
J1 (join1), J2, t1min, t2min, asc, dsc, and then the domed selectivity.
It is confusing because some of the parameters require highly modified
inverse logit transformation and some don't, so their values are not
intuitively obvious, and their descriptions in the User Manual are
misleading parameters 3 and 4 for example are not natural logs.
The six parameters are:

\if{html}{\out{<div class="sourceCode">}}\preformatted{1. the age/size where selectivity first becomes 1.0, p[1], a positive
value within the range of bin values

2. a modified inverse logit of the width of the plateau, typically 
-5 : 10, it is used to estimate the size/age where selectivity first 
begins to decline = peak2

3. the steepness of the ascending limb, a modified inverse logit, used 
when calculating t1min and asc

4. the steepness of the descending limb, a modified inverse logit, used 
when calculating t2min and dsc

5. the selectivity of the first age/size class, a modified inverse logit, 
used when calculating asc

6. the selectivity of the last age/size class,  a modified inverse logit, 
used when calculating dsc 

The descending limb of any dome shaped selectivity curves imply that the 
fishing gear used is unable to collect all representatives of the larger 
or older classes. The predicted numbers of smaller or younger animals, 
that are only partially selected, are inflated because of the partial 
selection. If any larger or older animals are, in fact, caught, then the 
same inflation can happen to those animals as a result of the partial 
selection implied by the dome shape. Small and young animals weigh 
very little, the same cannot be said for the larger or older animals. 
Some people refer to the extra biomass this phenomenon can imply as 
'ghost biomass', even though it might be real. Whatever the case, when 
using dome shaped selectivity it is best to be aware of this issue and 
to be cautious about how this is interpreted. The 20* terms in the J1 
and J2 factors are required to force the joins to be as effective as
required (see Methot and Wetzel 2013). 
}\if{html}{\out{</div>}}
}
\examples{
# from selex_length_example https://github.com/nmfs-ost/ss3-user-examples
  p <- c(45.8546,-3.18064,5.308,1.699,-999,0.75363)
  bins <- seq(11,99,2)
  sel <- domeSS3(p,bins)
  plot(bins,sel,type="l",xlab="Age",ylab="Selectivity",lwd=2)
}
\references{
Methot, R.D. and C.R, Wetzel (2013) Stock synthesis: A biological
and statistical framework for fish stock assessment and fishery management.
Supplementary material, Appendix A. Equs A1.30 - A1.34.
\emph{Fisheries Research} 142:86-99.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Hurtado-Ferro, F., Punt, A.E., and K.T. Hill (2014) Use of multiple 
selectivity patterns as a proxy for spatial structure. 
\\emph\{Fisheries Research\} 158:102-115.
}\if{html}{\out{</div>}}
}
