% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{filenametoPath}
\alias{filenametoPath}
\title{filenametoPath - safely add a filename or subdirectory to a path}
\usage{
filenametoPath(inpath, infile)
}
\arguments{
\item{inpath}{\itemize{
\item the path to be analysed
}}

\item{infile}{\itemize{
\item the filename or subdirectory to be added to the path 'inpath'
}}
}
\value{
the completed filename or extended path
}
\description{
filenametoPath - safely add a filename or subdirectory to a path
uses pathtype to get the seperator and then checks the end character.
If the separator is nothing of a '/' or a '//' then it reacts accordingly.
Without this one can unwittingly include extra separators or none at all.
}
\examples{
\dontrun{
resultpath <- "C:\\\\place\\\\Rcode\\\\place2\\\\dir3\\\\"
infile <- "starter.ss"
filenametoPath(resultpath,infile)
}
}
