% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{firstNum}
\alias{firstNum}
\title{firstNum - converts the first string in a vector to a number}
\usage{
firstNum(intxt)
}
\arguments{
\item{intxt}{\itemize{
\item a vector of characters, the first one of which is a number
}}
}
\value{
a single number
}
\description{
firstNum - converts the first string in a vector to a number.
If the first part of the string is not a number then firstNum will
issue a stop warning if the first part is not a number. If there are
tab markers present these will first be converted to spaces.
}
\examples{
\dontrun{
inputtxt <- "1  2   3   4"
x <- unlist(strsplit(inputtxt," "))
print(x)
print(firstNum(x))
}
}
