% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceoutput.r
\name{fixAgeCompVar}
\alias{fixAgeCompVar}
\title{fixAgeCompVar - in the ctl file adjust variances concerning age composition}
\usage{
fixAgeCompVar(incontrol, inrepout, fleets = c(1, 3), CAL = FALSE)
}
\arguments{
\item{incontrol}{\itemize{
\item a copy of the control file as stored in the 'analysis'
sub-directory and read in as text using readLines
}}

\item{inrepout}{\itemize{
\item the R object obtained from using SS_output, usually called
plotreport and saved as 'plotreport_', combined with the analysis names,
and then '.RData'
}}

\item{fleets}{\itemize{
\item identifies index of the fleet whose lengths are being corrected;
defaults to c(1,3)
}}

\item{CAL}{conditional Age-at-Length data rather than classical age composition
data; default = FALSE, meaning no conditional-age-at-length}
}
\value{
a list containing the text object containing a copy of the revised ctl file,
and the variance adjustment and the proportional change made.
}
\description{
fixAgeCompVar - in the ctl file adjust variances concerning age
composition. Do this by inputting a text copy of the control file and the report
object. The report  The report obtained from using SS_output, usually called
plotreport and saved as 'plotreport_', combined with the analysis names,
and then '.RData'. This is stored in the the sub-directory for the
particular analysis as the 'plotreport_analysis.RData' file.
The function finds the variance
adjustment suggested in the report file and transfers that into the
correct place in the control file. It uses the function
'SSMethod.TA1.8' to find the Francis weights required. This number has
to be multiplied by the current variance adjustment in the control file
and the result put back into the ctl file.
}
\examples{
print("See vignette for worked example.") # incontrol=control; inrepout=plotreport; fleets=agepos
}
