% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceoutput.r
\name{fixIndex}
\alias{fixIndex}
\title{fixIndex - adjusts the variances of the relative abundance indices}
\usage{
fixIndex(incontrol, inrepout, cepos = c(2, 3))
}
\arguments{
\item{incontrol}{\itemize{
\item a copy of the control file as stored in the 'analysis'
sub-directory and read in as text using readLines
}}

\item{inrepout}{\itemize{
\item the 'analysis.txt' file stored in the 'analysis' sub-directory and
read in as text using readLines
}}

\item{cepos}{\itemize{
\item the position of the indices of relative abundance in the sequence of
different data sources declared; defaults to c(2,3), meaning in the sequence of
TRAWL-CPUE-BIOMASS in teh data file, positions 2 and 3 refer to indices.
}}
}
\value{
a list containing the text object containing a copy of the revised ctl file,
and the variance adjustment and the proportional change made.
}
\description{
fixIndex - adjusts the variances of the relative abundance indices
It does this through using the readLines version of the control file and the
txt file from the analysis sub-directory. The function finds the variance
adjustments for numindex indices and adds them to the values already in the
ctl file.
}
\examples{
print("Still to make up an example for fixIndex")
}
