% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceoutput.r
\name{fixLengthVar}
\alias{fixLengthVar}
\title{fixLengthVar - adjust variances of length composition in the ctl file}
\usage{
fixLengthVar(incontrol, inrepout, fleets = c(1, 3, 4, 5))
}
\arguments{
\item{incontrol}{a copy of the control file as stored in the 'analysis'
sub-directory and read in as text using readLines}

\item{inrepout}{the 'analysis.txt' file stored in the 'analysis' sub-
directory and read in as text using readLines}

\item{fleets}{identifies index of the fleets whose lengths are being
corrected; defaults to c(1,3,4,5)}
}
\value{
a list containing the text object containing a copy of the revised
ctl file, and a matrix of the variance adjustments and the proportional
changes made.
}
\description{
fixLengthVar  in the ctl file adjust variances concerning length
composition. Do this by inputting a copy of the control file and the
report file. The report obtained from using SS_output, usually called
plotreport and saved as 'plotreport_', combined with the analysis names,
and then '.RData'. This is stored in
the the sub-directory for the particular analysis as the
'plotreport_analysis.RData' file.
The function finds the variance adjustment suggested in the report
file and transfers that into the correct place in the control file. It
searches in the #'     report file for the phrase 'Francis Weights -
len', which occurs more than once, but it uses the first occurrance. This
number has to be multiplied by the current variance in the control file
and the result put back into the ctl file.
}
\examples{
print("Still to make up an example")
}
