% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceoutput.r
\name{fixRecruit}
\alias{fixRecruit}
\title{fixRecruit - replaces the recruitment bias-adjustment in the ctl file}
\usage{
fixRecruit(incontrol, inrepout, sigRLim = c(0.25, 0.7))
}
\arguments{
\item{incontrol}{\itemize{
\item a copy of the control file as stored in the 'analysis'
sub-directory and read in as text using readLines
}}

\item{inrepout}{\itemize{
\item the R object obtained from using SS_output, usually called
plotreport and saved as 'plotreport_', combined with the analysis names,
and then '.RData'
}}

\item{sigRLim}{\itemize{
\item a vector or two numbers, the lower and upper limits for
sigmaR. This allows for adjustments to be made to sigmaR but only down to
the lower limit, which defaults to 0.25, or up to the upper limit, which
defaults to 0.7.
}}
}
\value{
a list containing the text object containing a copy of the revised ctl file,
and the variance adjustment and the proportional change made.
}
\description{
fixRecruit - replaces the recruitment bias-adjustment in the ctl
file. Do this by inputting a text copy of the control file and the report
file. The report file is the text generated after the model has been fitted
and the replist and plots are generated. This is stored in the the sub-directory
for the particular analysis as the 'analysis.txt' file. The function uses
'r4ss::SS_fitbiasramp' to find the suggested
varianceadjustment and transfers that into the
correct place in the control file.
}
\examples{
print("Still to make up an example for fixRecruit")
}
