% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{fixstarter}
\alias{fixstarter}
\title{fixstarter - saves a new starter.ss file ready to use the par file}
\usage{
fixstarter(directory, toscreen = FALSE)
}
\arguments{
\item{directory}{\itemize{
\item the calculation directory in which starter.ss can be found
}}

\item{toscreen}{\itemize{
\item defaults to FALSE, meaning no information is sent ot the
screen when this function is called.
}}
}
\value{
starter.ss is modified so that ss3 will use the previously
estimated par file as the starting point in the estimation
}
\description{
fixstarter - saves a new starter.ss file ready to use the
par file. This entails reading in the current starter.ss file, finding
the line with 'use init value' in it, changing the number at the start
from 0 to 1, and resaving the file.
}
\examples{
\dontrun{
print("An example has still to be written")
}
}
