% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{get_tuning_table}
\alias{get_tuning_table}
\title{get_tuning_table copied completely from r4ss currently not exported}
\usage{
get_tuning_table(
  replist,
  fleets,
  option,
  digits = 6,
  write = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{replist}{the output from SS_output}

\item{fleets}{which fleets to extract tuning variance information}

\item{option}{which tuning method to use, default = 'Francis', alternatives
include 'none', 'Francis', 'MI', or 'DI'}

\item{digits}{number of digits to round numbers to, default = 6}

\item{write}{Write suggested tunings to a file 'suggested_tunings.ss',
default = FALSE}

\item{verbose}{should comments and output be written to the console,
default = TRUE}
}
\value{
the tuning table
}
\description{
get_tuning_table is a function copied directly from r4ss. For
some reason, it is not exported from r4ss! So to use it in a comparison
one needs to add r4ss::: at its call, which upsets the building
process when compiling a package. This is a temporary fix and a request
will be made to the r4ss team to export the function (it seems like an
oversight).
}
\examples{
print("no example available")
}
