% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{getagelenkeys}
\alias{getagelenkeys}
\title{getagelenkeys extracts data from SS3 data file about age composition}
\usage{
getagelenkeys(inputdata, verbose = TRUE)
}
\arguments{
\item{inputdata}{the output of SS_readdat_3.30.}

\item{verbose}{default=TRUE, should messages be sent to the console.}
}
\value{
A list of all agelenkeys in the data set relating to gender, year,
and fleet, The number of scenarios, years, and fleets.
}
\description{
getagelenkeys is used to extract an sumry of all the age-
compositon data and conditional age-at-length data from an SS3 data file.
The inputdata are best generated using the SS_readdat_3.30. This function
has not been tested on versions of SS < 3.30.
}
\examples{
\dontrun{
  library(rforSS3)
  library(r4ss)
  library(codeutils)
  library(hplot)
  ddir <- getDBdir()
  wdir <- pathtopath(ddir,"/A_CodeR/SS3run/")
  store <- pathtopath(wdir,"egfish/")
  analysis <- "BaseCase"
  destination <- pathtopath(store,analysis)
  datfile <- pathtopath(destination,paste0(analysis,".dat"))
  dat <- SS_readdat_3.30(file=datfile,verbose = TRUE,
                         echoall = lifecycle::deprecated(),section = NULL)
  outscene <- getagelenkeys(dat)
  str(outscene,max.level=1)
  str(outscene$allscene,max.level=1)
}
}
