% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{getageprops}
\alias{getageprops}
\title{getaggprops calculates the proportions at age aggregated across years}
\usage{
getageprops(plotreport)
}
\arguments{
\item{plotreport}{the object output from SS_output}
}
\value{
a list of the proportional distribution of ages in each year, the
expected proportions of ages in each year, and the aggregated proportions
of ages across years
}
\description{
getaggprops calculates the proportions at age aggregated across years
}
\examples{
print("Wait n example data")
}
