% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{getalcomp}
\alias{getalcomp}
\title{getalcomp extracts any conditional age-at-length data from ss.dat}
\usage{
getalcomp(store, analysis)
}
\arguments{
\item{store}{this is the full path to the directory holding all the scenario
sub-directories.}

\item{analysis}{is a character string identifying the particular scenario
being explored. The dat file to be examined is held as 'analysis.dat'.}
}
\value{
two arrays of the female and make age-len key data as arrays of
length-bins x age-bins x years by gender plus the fleet names with
conditional age-at-length data. If no such data then NULL
}
\description{
getalcomp extracts any conditional age-at-length data directly
from the SS3 dat file. It produces an array of length-bin x age-bin X
year ready for eahc to be plotted.
}
\examples{
\dontrun{
  store <-  "c:/Users/userA/SS3/EG/afish/"
  analysis <- "sp_area_1"
  # inside sp_area_1 there needs to be a sp_area_1.dat file
  allcomp <- getalcomp(store-store,analysis=analysis)
}
}
