% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{getCase}
\alias{getCase}
\title{getCase - given a list of subdirectories this return the selected one}
\usage{
getCase(basecase, index = 1, printout = TRUE)
}
\arguments{
\item{basecase}{\itemize{
\item the vector of names of the brdiging analyses
}}

\item{index}{\itemize{
\item the index of the analysis wanted
}}

\item{printout}{\itemize{
\item a logical variable determining whether to print the name to the
screen or not; defaults to TRUE
}}
}
\value{
a character string containing the name of the subdirectory contianing the
ctl, dat, sta, for, and par files for the particular analysis wanted.
}
\description{
\itemize{
\item getCase - in a bridging analysis where one migrates an old assessment
to a new basecase one uses a 'basecase' directory to hold the subdirectories
chosen to hold each step in the bridging analysis. getCase simply holds this
vector of names and returns the selected one, optionally printing the name
to the screen. This functions main use is to simplify the use of the vector of
subdirectory names so that if one wants to add news one it stays simple.
}
}
\examples{
\dontrun{
basecase <- c("origbase24f","origbase24fb","origbase24z","newCatCE","newRecs",
"newRecRB","newAAL","newAgeComp","ageingerror","newLenComp",
"newLenatAge","balanced")
case <- 3
getCase(basecase,case)
}
}
