% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{getdatasets}
\alias{getdatasets}
\title{getdatasets tabulates the data in a printer friendly fashion}
\usage{
getdatasets(plotreport, store, analysis, verbose = TRUE)
}
\arguments{
\item{plotreport}{the output of SS_output}

\item{store}{the path to the directory within which the analysis is held}

\item{analysis}{the name of the SS3 model being explored}

\item{verbose}{Should warnings and progress be reported to the console?
default = TRUE.}
}
\value{
invisibly returns the catches by fleet in a matrix
}
\description{
getdatasets tabulates the data in a printer friendly fashion
so that this can be presented in appendices to enhance transparency.
Currently, it returns only the catches but this will be expanded in
later versions.
}
\examples{
# typical syntax
# getdatasets(plotreport=plotreport,store=store,analysis=analysis,
#             verbose=FALSE)
}
\seealso{
{
\link{SS_readdat_3.30}, \link{SS_output}
}
}
