% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{getfirst}
\alias{getfirst}
\title{getfirst - parse a given line to get a number occurring before #}
\usage{
getfirst(txtlist, index, rnd = TRUE)
}
\arguments{
\item{txtlist}{\itemize{
\item the results of applying readLines to the analysis.txt,
where 'analysis' is the name of the store directory being considered
}}

\item{index}{the location in the list of character strings to be parsed}

\item{rnd}{\itemize{
\item should the resulting number be truncated? Default = TRUE
}}
}
\value{
a single number deriving from analysis.txt used in the automatic
balancing of variances or recruitment residuals
}
\description{
getfirst - parse a single line to get the single number
at the front brealng the line up at '#'
}
\examples{
\dontrun{
print("being developed")
#getfirst(txt,2,rnd=FALSE)
}
}
