% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceoutput.r
\name{getfleets}
\alias{getfleets}
\title{getfleets - identifies the names given to all fleets in ctl file}
\usage{
getfleets(store, analysis)
}
\arguments{
\item{store}{the directory in which the seperate analysis directories
are kept}

\item{analysis}{the specific analysis within 'store' being
cosidered}
}
\value{
fleets - a vector of the fleet names
}
\description{
getfleets identifies the names given to all fleets as
listed in the ctl file. It uses the '\%' symbol to identify the
vector of fleet names so to use this requires that that symbol
is not used anywhere else in the ctl file, even in comments. This
will be fixed in later versions by excluding comment lines
}
\examples{
print("See the vignette for a worked example")
}
