% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.r
\name{getL50}
\alias{getL50}
\title{getL50 estimates the L50 from a selectivity vector}
\usage{
getL50(x)
}
\arguments{
\item{x}{a vector of selectivity values whose names are the lengths or ages
with which they are associated. The names are necessary for this function
to work.}
}
\value{
a single value defining the L50
}
\description{
getL50 searches for the value cloest to the L50 when given a
vector of selectivity at size or age
}
\examples{
print("Wait on an example")
}
