% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{getlast}
\alias{getlast}
\title{getlast - parse a given line to get a text occurring after #}
\usage{
getlast(txtlist, index)
}
\arguments{
\item{txtlist}{\itemize{
\item the results of applying readLines to the analysis.txt,
where 'analysis' is the name of the store directory being considered
}}

\item{index}{the location in the list of character strings to be parsed}
}
\value{
a character string starting with # to be added to the end of
the number obtained from getfirst, used in the automatic
balancing of variances or recruitment residuals
}
\description{
getlast - parse a single line to get the text occurring
after a '#'
}
\examples{
\dontrun{
print("being developed")
#getlast(txt,2)
}
}
