% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{getNum}
\alias{getNum}
\title{getNum - converts a given string in a vector to a number}
\usage{
getNum(intxt, index = 1)
}
\arguments{
\item{intxt}{\itemize{
\item a character vector containing the number wanted
}}

\item{index}{\itemize{
\item the index in the vector of character objects that
is to be converted to a number; defaults to 1
}}
}
\value{
a single number
}
\description{
getNum - converts a given string in a vector to a
number
}
\examples{
\dontrun{
inputtxt <- "[1]  2   3   4"
x <- unlist(strsplit(inputtxt," "))
print(x)
print(getNum(x,index=2))
}
}
