% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.r
\name{getplotreports}
\alias{getplotreports}
\title{getplotreports collets the plotreports and summaries into lists}
\usage{
getplotreports(store, subdir, picksubdir)
}
\arguments{
\item{store}{the directory containing the subdirectories of each of the
scenarios}

\item{subdir}{a vector of subdirectories to be found in store, obtained
using dir(store)}

\item{picksubdir}{a vector of the indicies of the subdirectories to include
into the comparisons.}
}
\value{
an invisible list containing a list of plotreports for each scenario
selected, as well as a matrix of summarySS3$answer vectors for each
scenario
}
\description{
getplotreports gathers the plotreports and the summarySS3
answers into a list containing a list of plotreports and a matrix of
summarySS3-answers. These can then be used when comparing the outputs
fromdifferent scenarios.
}
\examples{
# syntxt used
# replists <- getplotreports(store=store, subdir=subdir, picksubdir=c(3,6,7))
}
