% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{getprojdyn}
\alias{getprojdyn}
\title{getprojdyn tabulates the projected catches from a set of scenarios}
\usage{
getprojdyn(compscenes)
}
\arguments{
\item{compscenes}{the output from getreplist}
}
\value{
a list of projcatch, projdepl, projspawn, and parmscen - the
estimated parameters from each scenario
}
\description{
getprojdyn is used to tabulate the projected catches, spawning
biomass depletion, spawning biomass, and estimated parameters from a
set of different scenarios within SS3. It requires the use of the
function getreplists to aggregate the outputs from the different
scenarios.
}
\examples{
# compscenes <- getreplists(store=store,
#                           scenes=c("BC-5-5-1","BC-5-35-1","BC-5-5-75"),
#                           listname="plotreport")
# outproj <- getprojdyn(compscenes)
}
\seealso{
{
\link{getreplists}
}
}
