% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{getprops}
\alias{getprops}
\title{getprops gets proportions of agecomp or lencomp aggregated across years}
\usage{
getprops(compbase, fleetnames, comptype = "Len")
}
\arguments{
\item{compbase}{an object from SS_output output object, either
plotreport$agedbase or plotreport$lendbase}

\item{fleetnames}{the vector of fleetnames, usually plotreport$FlkeetNames}

\item{comptype}{default = 'Len' but could be 'Age', describes the typw of
data being extracted as length cor age omposition.}
}
\value{
a list of the proportional distribution of agecomps or lencomps in
each year, the related expected proportions in each year, and the
aggregated proportions across years
}
\description{
getprops gets proportions of agecomp or lencomp aggregated across years
}
\examples{
print("Wait n example data")
}
