% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{getStatus}
\alias{getStatus}
\title{getStatus - extract the values of B0 and Current Depletion}
\usage{
getStatus(txtlist)
}
\arguments{
\item{txtlist}{\itemize{
\item the list of character statements derived from
applyig readLines to the output file analysis.txt
}}
}
\value{
a vector of B0 and current Depltion
}
\description{
getStatus - extracts the values of B0 and Current
Depletion from the analysis.txt file readin using readLines.
}
\examples{
\dontrun{
print("run an example and apply this to the output file")
}
}
