% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{headtail}
\alias{headtail}
\title{headtail presents the head and tail of a matrix, data.frame, or tibble}
\usage{
headtail(x, n = 5, verbose = TRUE)
}
\arguments{
\item{x}{a matrix or data.frame}

\item{n}{how many rows from the first and last to be displayed, default=5}

\item{verbose}{should reports be made to the console}
}
\value{
a data.frame of 2 x n + 2 rows containing the head and tail of x
}
\description{
headtail is merely a utility function that displays both the
head and tail of a matrix or data.frame. It will translate a tibble into
a data.frame but any other object will return its structure using the
str(x, max.level=1) function. The head in the output is separated from
its tail by two lines of dots.
}
\examples{
dat <- matrix(rnorm(100,mean=10,sd=1),nrow=25,ncol=4,dimnames=list(1:25,1:4))
headtail(dat,5)
dat <- list(dat=dat,numrow=nrow(dat),numcol=ncol(dat))
headtail(dat,5)
}
