% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{pathend}
\alias{pathend}
\title{pathend - determines what character is at the end of a path}
\usage{
pathend(inpath)
}
\arguments{
\item{inpath}{\itemize{
\item the path to be analysed
}}
}
\value{
the end character of the path; either NA, '/', or '\\'
}
\description{
pathend - determines what character is at the end of a path
uses pathtype to get the seperator and then checks the end character
}
\examples{
\dontrun{
resultpath <- "C:\\\\place\\\\Rcode\\\\place2\\\\dir3\\\\"
pathend(resultpath)
}
}
