% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{pathtype}
\alias{pathtype}
\title{pathtype - finds the type of separator used in a path}
\usage{
pathtype(inpath)
}
\arguments{
\item{inpath}{\itemize{
\item the path to be analysed
}}
}
\value{
the type of path divider, either a 0 = '\\' or a
1 = '/'
}
\description{
pathtype - finds the type of separator used in a path;
this is either a '/' or a '\\'
}
\examples{
\dontrun{
resultpath <- "C:\\\\place\\\\Rcode\\\\place2\\\\dir3\\\\"
pathtype(resultpath)
}
}
