% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{plotagelenkey}
\alias{plotagelenkey}
\title{plotagelenkey plots all age-length keys derived form getagelenkeys}
\usage{
plotagelenkey(
  outcomp,
  rundir = "",
  plotscenes = NULL,
  pch = 1,
  pchcex = 1.25,
  pchcol = 2,
  console = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{outcomp}{the output of getagelenkeys applied to an SS3 data file,
which contains ageing data set up to apply conditional age-at-length.}

\item{rundir}{the directory into which to place plots if console=FALSE,
default = ''}

\item{plotscenes}{this can be used to plot sets of specific scenarios, made
up of gender + year + fleetname. default = NULL}

\item{pch}{the character to use in the plots, default = 1}

\item{pchcex}{the size of the character used in the plots}

\item{pchcol}{the colour of the character used in the plots}

\item{console}{should each plot go to the console, the default, or be
saved to rundir as a st of png files each identified by the scenes
plotted}

\item{verbose}{should any feedback and warnings be sent to the console.
default = TRUE}
}
\value{
nothing but it does generate a set of plots.
}
\description{
plotagelenkey is used to produce plots of all age-length keys
used in SS3 when using conditional age-at-length. The age-length keys
are naive in not having ageing error applied. That aspect remains under
development. The getagelenkeys functions identifies combinations of
the factors of gender, year, and fleet. The function is setup to plot
a maximum of eight plots in a single graph. If the number of scenarios
is greate than 8 then plotscenes can specify which to plot. Otherwise
it retains its default = NULL
}
\examples{
# outcomp=outscene; console=TRUE;plotscenes=NULL
# pchcex=1.25;pchcol=2;pch=1; verbose=TRUE
}
\seealso{
{
\link{getagelenkeys}
}
}
