% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{plotageprops}
\alias{plotageprops}
\title{plotageprops compares annual age composition data}
\usage{
plotageprops(
  agecomp1,
  agecomp2 = NULL,
  whichfleet = 1,
  height = 9,
  console = TRUE,
  rundir = "",
  scenarios = c(1, 2)
)
}
\arguments{
\item{agecomp1}{the output of getageprops applied to the first scenario's
plotreport}

\item{agecomp2}{the output of getageprops applied to the second scenario's
plotreport. Only the expected values are used from this one. Default
value = NULL, which means only the age proportions of the first agecomp1
will be plotted.}

\item{whichfleet}{the fleet number (actual names obtained from array dims)}

\item{height}{height of the plot, adjust depending on number of years/plots}

\item{console}{should the plot be saved as a png or go to console,
default = TRUE}

\item{rundir}{the directory into which any plot should be saved, default=''}

\item{scenarios}{the names of the scenarios, for labels.}
}
\value{
invisibly the filename of the plot for use with makehtml
}
\description{
plotageprops compares annual age composition data
by plotting each separately with the predicted proportions from just the
first two scenarios being compared, plotted on top of each other. The
proportions are scaled by effective sample size, not the nominal sample
size (see getageprops). This can be applied to two or one scenario.
}
\examples{
# compare==c("SGBC-5-4-100-43","SGBC-5-4-80-6")
}
\seealso{
{
\link{getageprops}, \link{makehtml}
}
}
