% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{plotaggage}
\alias{plotaggage}
\title{plotaggage plots age composition proportions aggregated across years}
\usage{
plotaggage(
  agg1,
  agg2 = NULL,
  whichfleet = 1,
  fleetname = "",
  height = 7,
  console = TRUE,
  rundir = "",
  scenarios = c(1, 2)
)
}
\arguments{
\item{agg1}{the output of getageprops applied to plotreport from scenario 1}

\item{agg2}{the output of getageprops applied to plotreport from scenario 1,
default = NULL, which would mean only the first scenario would be
plotted.}

\item{whichfleet}{the fleet number}

\item{fleetname}{the name of the fleet, default=''}

\item{height}{height of the plot, adjust as desired. default = 7}

\item{console}{should the plot be saved as a png or go to console,
default = TRUE}

\item{rundir}{the directory into which any plot should be saved, default=''}

\item{scenarios}{the names of the scenarios, for labels.}
}
\value{
invisibly the filename of the plot for use with makehtml
}
\description{
plotaggage uses the outputs of getageprops to plot a single
graph of the aggregated age proportions either singly or with the
predicted values of two scenarios plotted on top of each other
}
\examples{
# compare==c("SGBC-5-4-100-43","SGBC-5-4-80-6")
}
\seealso{
{
\link{getageprops}, \link{makehtml}
}
}
