% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{plotss3catches}
\alias{plotss3catches}
\title{plotss3catches generates a plot of catch by fleet and of total catch}
\usage{
plotss3catches(catches, extradir, analysis, console = TRUE)
}
\arguments{
\item{catches}{an output from getdatasets}

\item{extradir}{the full path to the directory fof extra plots and tables}

\item{analysis}{the name of the analysis or scenario whose catches are being
plotted}

\item{console}{should the plot go to the console or be saved as a png file
to the extradir}
}
\value{
nothing but it does plot a graph
}
\description{
plotss3catches generates a two-panel plot with catches-by-fleet
and total catches across all fleets, which also has a loess smoother
imposed to assist in identifying any trends in total catch.
}
\examples{
# typical syntax
# plotss3catches(catches=catches, extradir=extradir,console=FALSE)
}
