% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.r
\name{profilestarter}
\alias{profilestarter}
\title{profilestarter gets the starter file ready for a likelihood-profile}
\usage{
profilestarter(
  calc,
  findtext = c("init_values_src", "prior_like"),
  verbose = FALSE
)
}
\arguments{
\item{calc}{the directory defined as the one in which the SS3 calculaitons
will occur}

\item{findtext}{a vector containing identifying text for the required two
lines in the starter file. If they are not c("init_values_src",
"prior_like"), then, obviously, you should alter 'findtext' to reflect
whatever text you have put there.}

\item{verbose}{should progress be sent to the console? default = FALSE}
}
\value{
nothing but it does alter the contents of the starter.ss file
}
\description{
profilestarter reads in the starter.ss file from the calc
directory - ensure that is the starter.ss for the scenario whose
likelihood profile you want. It then changes the value of the two fields
init-values-src and prior-like to 1 rather than zero. If they are already
set to 1 this will remain the case.
}
\examples{
print("Wait on an example being prepared")
}
