% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{projectedcatches}
\alias{projectedcatches}
\title{projectedcatches allows alternative SS3 model catches to be compared}
\usage{
projectedcatches(plotreport)
}
\arguments{
\item{plotreport}{the output from SS_outout, a replist}
}
\value{
A matrix of the catches including the projections, of spawnB,
recruits, depletion, total catches and ERA
}
\description{
projectedcatches After fitting an optimal stock assessment
model using SS3 it is usual to generate the predicted catches needed to
move a fishery towards a given harvest strategy policy target reference
point. To do this it used to be standard practice to use recruitment
levels taken from the predicted stock recruitment relationship. However,
it has become a good practice to consider the recent history of
predicted recruitment and modify the projected recruitment levels taken
of the stock recruitment relationship such that they are similar to the
last 5 to 10 years. SS3 has an option in the forecast file called either
# multiplier on base recruitment or # scalar/multiplier or value applied
to SR-curve. If such a modification is made it would be usual practice
to compare the predicted catches with and with the adjustment. This
function extracts the total catches
}
\examples{
# catches <- projectedcatches(plotreport=paste0(destination,"filename.Rdata"))
}
\seealso{
{
\link[r4ss]{SS_output}, \link{getreplists}, \link{projreceffects}
}
}
