% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{runSS3}
\alias{runSS3}
\title{runSS3 - calls SS3 -nohess twice, then SS3 to calculate the hessian}
\usage{
runSS3(wkdir, exec = "SS3", calcdir = "calc/")
}
\arguments{
\item{wkdir}{\itemize{
\item the full working directory eg fld2016
}}

\item{exec}{the name of the executable within the calc subdirectory to use
with the particular set of SS files}

\item{calcdir}{\itemize{
\item the directory in which all calculations occur; defaults to
"calc/"
}}
}
\value{
conducts three separate SS3 runs for the input data. Sends messages
to the screen for each run. Generates numerous files in teh /calc directory.
}
\description{
runSS3 - this switches to /calc, sets up and calls SS3 -nohess,
then alters starter.ss to use the estimated .par file from the first run.
It recalls SS3 -nohess. Finally, it re-uses the new .par file in a call to
SS3, which also estimates the Hessian matrix. Finally it returns to the
working directory.
}
\examples{
\dontrun{
print("Still need to develop a real example using included datasets")
}
}
