% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{sel24}
\alias{sel24}
\title{sel24 implements SS3's selectivity pattern 24}
\usage{
sel24(p, L)
}
\arguments{
\item{p}{a vector of six parameters.}

\item{L}{a vector of the mean of nL age/size classes}
}
\value{
a vector of selectivity at length/age
}
\description{
sel24 uses 6 parameters and a set of mean size or age classes
to calculate SS3's selectivity pattern 24, which is a double normal with
a defined peak at 1.0, and tail controls, that is it has parameters for
the selectivity of the initial and final size/age classes. There is an
ascending limb and a descending limb with the potential of a plateau in
between. The six parameters are 1) the age/size where selectivity first
becomes 1.0, 2) the size/age where selectivity first begins to decline,
3) the steepness or width of the ascending limb, 4) the steepness or
width of the descending limb, 5) the selectivity of the first age/size
class, and 6) the selectivity of the last age/size class. The descending
limb of any dome shaped selectivity curves imply that the fishing gear
used is unable to collect all representatives of the larger or older
classes. The predicted numbers of smaller or younger animals, that are
only partially selected, are inflated because of the partial selection.
If any larger or older animals are, in fact, caught, then the same
inflation can happen to those animals as a result of the partial
selection implied by the dome shape. Small and young animals weigh very
little, the same cannot be said for the larger or older animals. Some
people refer to the extra biomass this phenomenon can imply as 'ghost
biomass', even though it might be real. Whatever the case, when using
dome shaped selectivity it is best to be aware of this issue and to be
cautious about how this is interpreted. A version of this function was
first developed for the MQMF package (Haddon, 2023).
}
\examples{
  L <- seq(180,410,1)
  p <- p <- c(230,250,155,600,-5,1.5)
  sel <- sel24(p,L)
  plot(L,sel,type="l",xlab="Age",ylab="Selectivity",lwd=2)
}
\references{
Methot, R.D. and C.R, Wetzel (2013) Stock synthesis: A biological
and statistical framework for fish stock assessment and fishery
management. Supplementary material, Appendix A. Equs A1.30 onwards.
\emph{Fisheries Research} 142:86-99.

Haddon M (2023). \emph{MQMF: Modelling and Quantitative Methods
in Fisheries}. R package version 0.1.5,\url{https://github.com/haddonm/MQMF/}.
}
