% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.r
\name{sscopyto}
\alias{sscopyto}
\title{sscopyto copies the ctl, dat, par, sta, and for files to new directory}
\usage{
sscopyto(origin, fromdir, todir, neworigin = NA)
}
\arguments{
\item{origin}{the store directory in which the fromdir subdirectory exists}

\item{fromdir}{the name of the source sub-directory}

\item{todir}{the full path to the destination sub-directory}

\item{neworigin}{the new store directory, Default = NA. If it is NA then
the destination store directory will be the same as the origin store. If
neworigin has a full pathway in it, then it will be used instead of the
origin and the todir will be created in it if it does not exist.
neworigin might be akin to the 'basecase' store but be called
'profile' or 'sensitivity'}
}
\value{
a vector of 1 or -1 denoting which files are transferred
}
\description{
sscopyto copies the ctl, dat, par, sta, and for files from one
directory to another, changing the name to match to destination
sub-directory as appropriate. Used in conjunction with rforSS3 in the
development of a bridging analyis, or to prepare for a likelihood
profile, or sensitivity analysis. Now 'copyto' includes the option of
copying to a completely different path and will create the 'todir' if it
does not exist.
}
\examples{
\dontrun{
# When conducting a bridging analysis using 'store' to define the path to
# the various subdirecotries holding each step
sscopyto(store,"addcatches","addsurvey")
#
# When the bridging analysis is completed and one might want to do a
# set of likelihood profiles
sscopyto(store,"basecase17","profileM",
       neworigin="C:/Rcode/ss3/oro2017/profile/")
# is equivalent to:
sscopyto(origin=store,fromdir="basecase17",todir="profileM",
       neworigin="C:/Rcode/ss3/oro2017/profile/")
}
}
