% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{storeresults}
\alias{storeresults}
\title{storeresults - Copies result files back into destination directory}
\usage{
storeresults(origin, destination)
}
\arguments{
\item{origin}{\itemize{
\item the directory from which the files are t obe copied.
}}

\item{destination}{\itemize{
\item the directory into which the files are to be copied.
}}
}
\value{
The files listed under description are copied to the destination
directory. If any are missing a warning is given
}
\description{
storeresults - Stores result files back into destination
directory; the origin would usually be 'calc'and the destination would
generally be the directory from which the SS3 files were copied. The
files copied are: checkup.sso, CompReport.sso, covar.sso, CumReport.sso,
echoinput.sso, Forecast-report.sso, ParmTrace.sso, Report.sso,
SIS_table.sso, warning.sso, ss3.ctl, ss3.dat, ss3.par, ss3.rep, ss3.cor,
ss3.std, forecast.ss, starter.ss, 1rundetails.txt, 2rundetails.txt,
3rundetails.txt
}
\examples{
\dontrun{
print("An example has still to be written")
}
}
