% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{summarySS3}
\alias{summarySS3}
\title{summarySS3 pulls out summary information from report file}
\usage{
summarySS3(outfile)
}
\arguments{
\item{outfile}{\itemize{
\item text file containing screen dump from SS_output and SS_plots
}}
}
\value{
a list containing run details, number of paramters, and the
likelihoods estimated
}
\description{
summarySS3 pulls out summary information from report file
}
\examples{
\dontrun{
print("After using runSS3 on an 'analysis' this will generate an")
print("'analysis.txt' file in the 'analysis' sub-directory.")
print("summarySS3('analysis.txt')")
}
}
