% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{plotagecomp}
\alias{plotagecomp}
\title{plotagecomp literally plots the conditional age-at-length data}
\usage{
plotagecomp(
  compsex,
  gender,
  fleet,
  rescale = 0,
  console = TRUE,
  plotdir = "",
  plotout = TRUE
)
}
\arguments{
\item{compsex}{the array of lengthbin x agebins x year for a given gender}

\item{gender}{a character string denoting the gender being illustrated}

\item{fleet}{which fleet is the data representing, a character string}

\item{rescale}{if the default value fails to provide a statisfactory plot
thenm enter a value here to obtain what's needed. The default = 0, in
which case the number of age-bins is divided by 18, which usally works.}

\item{console}{should the plot go to the console or to a file. default=TRUE}

\item{plotdir}{the full path to the plotting directory if a file is to be
saved.}

\item{plotout}{should a plot be produced or just output the alcomp? defualt
= TRUE}
}
\value{
invisibly a list of the filename used and a copy of the 3D array
of age-length-year developed.
}
\description{
plotagecomp generates a plot of the conditional age-at-length
data where each data point is jittered so that a direct impression of
the density of observations is presented. More than 20 plots in a single
panel would get too busy, it is up to the user to control the flow of
the number of years of data sent to the function.
}
\examples{
print("wait on example data - wait a while!")
# compsex=compfem[pickflt,]; gender="Female"; fleet=fleets[1]; rescale=0.0;
# console=TRUE;plotdir="";plotout=TRUE 
}
\seealso{
{
\link{plotagelen}, \link{getalcomp}
}
}
