% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{plotagelen}
\alias{plotagelen}
\title{plotagelen plots a jittered version of conditional age-at-length data}
\usage{
plotagelen(alcomp, gender, fleet, console = TRUE, plotdir = "")
}
\arguments{
\item{alcomp}{the output from plotagecomp an array of length x age x year}

\item{gender}{a character label for the plots, usually 'Male', 'Female', or
'mixed'}

\item{fleet}{which fleet, again for a plot label, either a number of name}

\item{console}{should the plot go to the console or a file, default=TRUE}

\item{plotdir}{the full path to the directory in which to store the file,
default = ''}
}
\value{
nothing but it does generate a plot
}
\description{
plotagelen uses base::jitter to plot any conditional
age-at-length observed data for each year. r4ss has replaced the usual
jitter function hence the need for the namespace.
}
\examples{
print("Wait on data sets")
# alcomp=alfem;gender="Female";fleet=fleets[1];console=TRUE;plotdir=""
}
\seealso{
{
\link{plotagecomp}, \link{getalcomp}
}
}
