% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.r
\name{plotselex}
\alias{plotselex}
\title{plotselex generates a plot of the selectivity used in SS3}
\usage{
plotselex(
  plotreport,
  sex = "Female",
  yrs = c(1984, 2004, 2016),
  upbound = 0,
  legendcex = 1.1,
  labelcex = 1.5
)
}
\arguments{
\item{plotreport}{the list object output from teh SS_output function after
running an SS3 scenario}

\item{sex}{which gender should be plotted (should eb Female or Male)}

\item{yrs}{a single or vector of years to be plotted}

\item{upbound}{the upper bound of the plot, which can be used to remove
extra or unwanted size classes that are all set to 1.0}

\item{legendcex}{the font size fot eh legend default = 1.1}

\item{labelcex}{the font size for the labels and legend default 1.5}
}
\value{
It generates a plot and returns, invisibly, a vector of L50 values
for each estimated selectivity curve
}
\description{
plotselex provides an alternative plot of the selectivity
curves for a given gender and for specific years. This is required
because the 3-D plots generated by R4SS are often difficult to interpret.
}
\examples{
# plotreport=plotreport; sex="Female";yrs=c(1984,2004,2016); upbound=365
}
