% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{plotselex}
\alias{plotselex}
\title{plotselex generates a plot of the selectivity used in SS3}
\usage{
plotselex(
  plotreport,
  sex = "Female",
  upbound = 0,
  legendcex = 1.1,
  labelcex = 1.5,
  console = TRUE,
  rundir = ""
)
}
\arguments{
\item{plotreport}{the output from SS_output after running SS3}

\item{sex}{the gender as in 'Male' or 'Female', default = 'Female', use
this even if sex = 0 is used, it will be ignored}

\item{upbound}{upper value of ylim, default = 0, meaning ylim=c(0,1.1)}

\item{legendcex}{the font size for the legend, default = 1.1}

\item{labelcex}{default = 1.5, font size for axis labels}

\item{console}{default = TRUE, should plot go to console or png file?}

\item{rundir}{directory into which to place plots, default=""}
}
\value{
invisibly the filename of the plot being made.
}
\description{
plotselex provides an alternative plot of the selectivity
curves for a given gender and for specific years. This is required
because the 3-D plots generated by R4SS are often difficult to interpret.
By inputting specific years plotselex can address selectivity in
different time-blocks.
}
\examples{
# plotreport=plotreport; sex="Female";yrs=c(1984,2004,2016); upbound=0; console=TRUE
}
