% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3_utils.r
\name{storeresults}
\alias{storeresults}
\title{storeresults - Copies result files back into destination directory}
\usage{
storeresults(
  origin,
  destination,
  replaceCS = FALSE,
  getfiles = c("CompReport.sso", "covar.sso", "CumReport.sso", "echoinput.sso",
    "Forecast-report.sso", "Report.sso", "SIS_table.sso", "warning.sso", "ss3.par",
    "ss3.std", "control.ss_new", "starter.ss_new")
)
}
\arguments{
\item{origin}{\itemize{
\item the directory from which the files are to be copied.
}}

\item{destination}{\itemize{
\item the directory into which the files are to be copied.
}}

\item{replaceCS}{should the control and starter files be replaced by the
SS3 generated control.ss_new and starter.ss_new files (this is useful
if using the tune_comps function for iterstive re-weighting)}

\item{getfiles}{a vector of filenames to be retrieved from the origin
directory (usually calc) and copied to the destination directory. The
current files chosen are: CompReport.sso, covar.sso, CumReport.sso,
echoinput.sso, Forecast-report.sso, ParmTrace.sso Report.sso,
SIS_table.sso, warning.sso, ss3.par, ss3.std, control.ss_new, and
starter.ss_new}
}
\value{
The files listed under getfiles are copied from the origin to the
destination directory. If any are missing a warning is given
}
\description{
storeresults - Stores result files back into destination
directory; the origin would usually be 'calc'and the destination would
generally be the directory from which the SS3 files were copied.
}
\examples{
\dontrun{
# typical syntax
 store <- pathtopath(wkdir,"basecase/")  # define directories
 calc <- pathtopath(wkdir,"calc/")
 
 item <- 7    #  which scenario subdirectory inside basecase to be run
 getCase(index=item,basecase) # this lists the basecase indices to the screen
 analysis <- getCase(index=item,basecase)  # define analysis directory
 destination <- pathtopath(store,analysis)
 # run SS3
 storeresults(calc,destination,replaceCS=FALSE)
}
}
\seealso{
{
\link{tune_comps}
}
}
