% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.r, R/ss3_utils.r
\name{summarizeSS3}
\alias{summarizeSS3}
\title{summarizeSS3 provides a set of summary statistics from SS_output}
\usage{
summarizeSS3(replist)

summarizeSS3(replist)
}
\arguments{
\item{replist}{the object generated by SS_output}
}
\value{
a list of results and estimated parameters, and all likelihoods

a list of results and estimated parameters, and all likelihoods
}
\description{
summarizeSS3 provides a set of summary statistics from the
object generated by SS_output. The vector 'answer' contains the final
year, the final depletion, Bzero, 1-SPR, the total likelihood, and other
likelihoods. In addition, a matrix of parameters are output which
contains all estimated parameters = phase > 0 (including the recruitment
deviates)

summarizeSS3 provides a set of summary statistics from the
object generated by SS_output. The vector 'answer' contains the final
year, the final depletion, Bzero, 1-SPR, the total likelihood, and other
likelihoods. In addition, a matrix of parameters are output which
contains all estimated parameters = phase > 0 (including the recruitment
deviates)
}
\examples{
print("See the vignette for a worked example")
print("See the vignette for a worked example")
}
